requirements.txt
# Project Garuda — Dependencies
# Core
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
requests>=2.28.0

# Tracking
deep-sort-realtime>=1.3.2

# Logging & CLI
colorama>=0.4.6

# Testing
pytest>=7.4.0

# Export & Optimization (optional — install as needed)
# onnx>=1.14.0
# onnxruntime>=1.15.0
# onnxruntime-gpu>=1.15.0       # For Jetson/CUDA
# tflite-runtime>=2.13.0        # For Raspberry Pi (ARM)
# tensorrt>=8.6.0               # For NVIDIA Jetson TensorRT
# openvino>=2023.0              # For Intel devices

# Development (optional)
# black>=23.0
# flake8>=6.0
# mypy>=1.5
